+++ bricks/page/__brick__/lib/features/{{feature_name.snakeCase()}}/presentation/viewmodels/{{page_name.snakeCase()}}_viewmodel.dart
@@ -1,17 +1,12 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-/// State for {{page_name.pascalCase()}}
-class {{page_name.pascalCase()}}State {
-  // TODO: Add state fields
-  const {{page_name.pascalCase()}}State();
-}
+import '{{page_name.snakeCase()}}_state.dart';
 
 /// ViewModel for {{page_name.pascalCase()}}Page
 class {{page_name.pascalCase()}}ViewModel extends Notifier<{{page_name.pascalCase()}}State> {
   @override
   {{page_name.pascalCase()}}State build() {
     // TODO: Initial state logic
-    return const {{page_name.pascalCase()}}State();
+    return const {{page_name.pascalCase()}}State.loaded();
   }
 
   // TODO: Add methods for business logic
@@ -21,4 +16,3 @@ class {{page_name.pascalCase()}}ViewModel extends Notifier<{{page_name.pascalCas
 final {{page_name.camelCase()}}ViewModelProvider = NotifierProvider<{{page_name.pascalCase()}}ViewModel, {{page_name.pascalCase()}}State>(
   {{page_name.pascalCase()}}ViewModel.new,
 );
-